Start of working with placeables, currently failed attempt at moving out timers
I'm making Building into a Node, that's a Placeable (which will eventually be a component of a tile), but that's barely started. Instead, the goal was to move the timer for the grill out of the UI into the building code, but it's not working no matter how I try it (attaching it in the scene tree, doing it entirely in code)--the current code works, and there's just a lot of code that's duplicated in the test scene and building atm.