Timer (mostly) decoupled
I couldn't launch a timer from Building directly, so I had it send a signal to start a timer somewhere, so even though the timer is still in the big scene it could much more easily be moved now (likely to tile eventually). Now to do the same things with Field before moving on to new things!