Placing things in Build Mode complete, initial work on activating tiles
Can now go into build mode, click on something to build, click on a tile, and it'll be attached to that tile (and the tile will change visually, as my first use of those textures). Also starting to work with clicking tiles outside of build mode, with basic placeholders (and work with overriding functions).