Initial, very buggy take on Worker paths
The basics of worker paths are in: you can place things on the grid, make a worker, and assign it start and end points and it'll (try to) make a path that only goes on roads and buildings. However, not only does it not do anything yet, the path generation is buggy--some short paths seem to work, but longer paths seem too long, and others won't generate at all. This will take a while to debug, since I need to make sense of the paths.