Main work on Cutscenes complete
Lots of stuff (that should have been committed in smaller batches, but the first cutscene is done in basic form, including a lot of supporting resources, the scripts to make the next button work, and even some graphics (and the first resembling final--thanks LKB). There's still more (the previous messages popup and general styling, including a background), but it's a great start.