Currency implemented
As expected, currency was simple, adding a few functions to Inventory and support to the various UI's and Resources. There's a bug with the coin count not updating on the TileTest UI right away for some reason, and I want to add the costs to the buttons, but that can be the next commit.