Initial TilePopup work
The final (non-worker) functional piece of the level scene has been implemented, and while it's a CanvasLayer now, it's still in one place--though it fully replicates the test scene's functionality. I want to have it dynamically move to the tile you select, and then also explore using a progress bar to fix the lack of feedback in the test scene (and there are pieces, like putting a current_tile in the tile popup).