Level 1 End Cutscene done, Level 1 Play simplified
After playtesting, the first level was a little complicated and long, so I removed pancakes from it (which led to the DSD item becoming a plot point later). I also added the cutscene for the end of Level 1--Level 2's start cutscene is written and coming next, as is Level 2 proper (minus some new resources). I also need to make instructions on the title screen, and likely a cookbook as well.