Level 3 started, almost all art/resources complete
Lots of resources/art added as Level 3 is startable--all it needs is an order list, as well as code to handle ending immediately when the Chicken Dessert is fulfilled. I also need code for boosting when trashing the Mystery Ketchup. Start cutscene needs to be ported and End cutscene needs to be finished, then the major hurdle left is art: the big backgrounds, the bad Taxxie, the blowing Glerry, and the Customer Area. Otherwise there's just minor things like the in-level cookbook (with pausing the order timer) and actually setting a loss condition (and the options).